requirements.txt
fastapi
pydantic
uvicorn
python-dotenv
pymongo[srv]==3.12
loguru
mongoengine
